+++ packages/yourloops/lib/utils.ts
@@ -31,7 +31,7 @@ import _ from "lodash";
 import { Units } from "../models/generic";
 import { IUser, Settings } from "../models/shoreline";
 import httpStatus from "./http-status-codes";
-import { t } from "./language";
+import { t } from "./locales";
 import metrics from "./metrics";
 
 // From https://emailregex.com/